pyproject.toml
[project]
name = "curator"
version = "0.1.0"
description = "Content acquisition and curation service"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "apscheduler>=3.10.0",
    "yt-dlp>=2024.1.0",
    "feedparser>=6.0.0",
    "readability-lxml>=0.8.0",
    "structlog>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
curator = "curator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/curator"]
